+++ kedro_airflow_k8s/template.py
@@ -3,10 +3,21 @@ from pathlib import Path
 from typing import Optional
 
 import jinja2
+import kedro
 from jinja2.environment import TemplateStream
 from slugify import slugify
 
 from kedro_airflow_k8s import version
+
+
+def _get_mlflow_url(context_helper):
+    try:
+        import importlib
+
+        importlib.import_module("kedro_mlflow")
+        return context_helper.mlflow_config["mlflow_tracking_uri"]
+    except (ModuleNotFoundError, kedro.config.config.MissingConfigException):
+        return None
 
 
 def _get_jinja_template():
@@ -49,13 +60,13 @@ def _create_template_stream(
         base_nodes=nodes_with_no_deps,
         bottom_nodes=bottom_nodes,
         config=context_helper.config,
+        mlflow_url=_get_mlflow_url(context_helper),
         env=context_helper.env,
         project_name=context_helper.project_name,
         dag_name=dag_name,
         image=image,
         schedule_interval=schedule_interval,
         git_info=context_helper.session.store["git"],
-        mlflow_url=context_helper.mlflow_config["mlflow_tracking_uri"],
         kedro_airflow_k8s_version=version,
         include_start_mlflow_experiment_operator=(
             Path(__file__).parent / "operators/start_mlflow_experiment.py"