+++ kedro_airflow_k8s/cli.py
@@ -8,7 +8,10 @@ from tabulate import tabulate
 
 from kedro_airflow_k8s.airflow import AirflowClient
 from kedro_airflow_k8s.context_helper import ContextHelper
-from kedro_airflow_k8s.template import get_dag_filename_and_template_stream
+from kedro_airflow_k8s.template import (
+    get_cron_expression,
+    get_dag_filename_and_template_stream,
+)
 
 
 @click.group("airflow-k8s")
@@ -44,7 +47,7 @@ def airflow_group(ctx, metadata, env):
 def compile(ctx, image, target_path="dags/"):
     """Create an Airflow DAG for a project"""
     dag_filename, template_stream = get_dag_filename_and_template_stream(
-        ctx, image=image
+        ctx, image=image, cron_expression=get_cron_expression(ctx)
     )
 
     target_path = Path(target_path) / dag_filename
@@ -59,7 +62,7 @@ def compile(ctx, image, target_path="dags/"):
     "--output",
     "output",
     type=str,
-    required=True,
+    required=False,
     help="Location where DAG file should be uploaded, for GCS use gs:// or "
     "gcs:// prefix, other notations indicate locally mounted filesystem",
 )
@@ -77,9 +80,10 @@ def upload_pipeline(ctx, output: str, image: str):
     Uploads pipeline to Airflow DAG location
     """
     dag_filename, template_stream = get_dag_filename_and_template_stream(
-        ctx, image=image
+        ctx, image=image, cron_expression=get_cron_expression(ctx)
     )
 
+    output = output or ctx.obj["context_helper"].config.output
     with fsspec.open(f"{output}/{dag_filename}", "wt") as f:
         template_stream.dump(f)
 
@@ -90,6 +94,7 @@ def upload_pipeline(ctx, output: str, image: str):
     "--output",
     "output",
     type=str,
+    required=False,
     help="Location where DAG file should be uploaded, for GCS use gs:// or "
     "gcs:// prefix, other notations indicate locally mounted filesystem",
 )
@@ -98,7 +103,7 @@ def upload_pipeline(ctx, output: str, image: str):
     "--cron-expression",
     type=str,
     help="Cron expression for recurring run",
-    required=True,
+    required=False,
 )
 @click.pass_context
 def schedule(ctx, output: str, cron_expression: str):
@@ -106,9 +111,10 @@ def schedule(ctx, output: str, cron_expression: str):
     Uploads pipeline to Airflow with given schedule
     """
     dag_filename, template_stream = get_dag_filename_and_template_stream(
-        ctx, cron_expression
+        ctx, cron_expression=get_cron_expression(ctx, cron_expression)
     )
 
+    output = output or ctx.obj["context_helper"].config.output
     with fsspec.open(f"{output}/{dag_filename}", "wt") as f:
         template_stream.dump(f)
 
@@ -119,7 +125,7 @@ def schedule(ctx, output: str, cron_expression: str):
     "--output",
     "output",
     type=str,
-    required=True,
+    required=False,
     help="Location where DAG file should be uploaded, for GCS use gs:// or "
     "gcs:// prefix, other notations indicate locally mounted filesystem",
 )
@@ -151,7 +157,7 @@ def schedule(ctx, output: str, cron_expression: str):
 @click.pass_context
 def run_once(
     ctx,
-    output: str,
+    output: Optional[str],
     dag_name: Optional[str],
     wait_for_completion: Optional[int],
     image: Optional[str],
@@ -160,18 +166,17 @@ def run_once(
     Uploads pipeline to Airflow and runs once
     """
     dag_filename, template_stream = get_dag_filename_and_template_stream(
-        ctx, dag_name=dag_name, image=image
+        ctx, dag_name=dag_name, image=image, cron_expression=None
     )
     context_helper = ctx.obj["context_helper"]
+    output = output or context_helper.config.output
 
     with fsspec.open(f"{output}/{dag_filename}", "wt") as f:
         template_stream.dump(f)
 
-    airflow_client = AirflowClient(
-        context_helper.airflow_config["airflow_uri"]
-    )
+    airflow_client = AirflowClient(context_helper.config.host)
     dag = airflow_client.wait_for_dag(
-        dag_id=dag_name or context_helper.context.package_name,
+        dag_id=dag_name or context_helper.config.run_config.run_name,
         tag=f'commit_sha:{context_helper.session.store["git"]["commit_sha"]}',
     )
     dag_run_id = airflow_client.trigger_dag_run(dag.dag_id)
@@ -193,9 +198,7 @@ def list_pipelines(ctx):
     """
 
     context_helper = ctx.obj["context_helper"]
-    airflow_client = AirflowClient(
-        context_helper.airflow_config["airflow_uri"]
-    )
+    airflow_client = AirflowClient(context_helper.config.host)
 
     dags = airflow_client.list_dags("generated_with_kedro_airflow_k8s")
 
@@ -222,8 +225,7 @@ def list_pipelines(ctx):
 @click.pass_context
 def ui(ctx, dag_name: Optional[str] = None):
     """Open Apache Airflow UI in new browser tab"""
-    context_helper = ctx.obj["context_helper"]
-    host = context_helper.airflow_config["airflow_uri"]
+    host = ctx.obj["context_helper"].config.host
     if dag_name:
         host = f"{host}/tree?dag_id={dag_name}"
     webbrowser.open_new_tab(host)