+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 -   Command `list-pipelines` allows display of pipelines which were generated with the plugin
 -   Command `ui` opens webbrowser with Apache Airflow
--   Move options to config file
+-   Move options to config file and restructure
 
 ## [0.1.2] - 2021-03-24