+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 -   Command `list-pipelines` allows display of pipelines which were generated with the plugin
+-   `MLflow` enabled only if `kedro-mlflow` present in dependencies and configuration is in place 
 
 ## [0.1.2] - 2021-03-24